Move form logging tests to shared infrastructure
